refactor - move helper methods to base
